Move sphere_coords to a static method of scattering theory
