+++ holopy/scattering/tests/test_mie.py
@@ -35,7 +35,7 @@ from holopy.scattering.scatterer import (
 from holopy.scattering.theory import Mie
 from holopy.scattering.errors import TheoryNotCompatibleError, InvalidScatterer
 from holopy.core.metadata import (
-    detector_grid, detector_points, to_vector, sphere_coords, update_metadata)
+    detector_grid, detector_points, to_vector, update_metadata)
 from holopy.core.process import subimage
 from holopy.scattering.tests.common import (
     sphere, xschema, scaling_alpha, yschema, xpolarization, ypolarization,
@@ -341,7 +341,7 @@ def test_raw_fields():
     pol = to_vector((0, 1))
     sch = detector_grid(3, .1)
     wavevec=2*np.pi/(wavelen/index)
-    pos = sphere_coords(sch, (10, 10, 5), wavevec=wavevec)
+    pos = Mie.sphere_coords(sch, (10, 10, 5), wavevec=wavevec)
     rf = Mie()._raw_fields(np.vstack((pos['r'], pos['theta'], pos['phi'])), sp, medium_wavevec=wavevec, medium_index=index, illum_polarization=pol)
     assert_allclose(rf, [[(0.0015606995428858754-0.0019143174710834162j),
   (-0.0003949071974815011-0.0024154494284017187j),
